Movie Lens Project
Project Data Description: MovieLens data sets were collected by the GroupLens Research Project at the University of Minnesota.
Context:
The GroupLens Research Project is a research group in the Department of Computer Science and Engineering at the University of Minnesota. The data is widely used for collaborative filtering and other filtering solutions. However, we will be using this data to act as a means to demonstrate our skill in using Python to “play” with data.